Move the frontend handlers out of the app.py
